Handle multiple source archives
This is really just to enable conversion for `pacemaker`, which has two
archives to be downloaded and unpacked.

This implementation assumes that the number of archives stored in the
lookaside cache correspond to the number of sources referenced in the
SPEC-file, though this not might be true in all the cases.

Signed-off-by: Hunor Csomortáni <csomh@redhat.com>