Apply patches as commits
This is still a very hacky solution, as patchnum is just a sequence
number and we have no reference what the patch file was called from
which the commit is created. But to get all those, we'll probably need
to patch rpm itself, as all this information seems to be only available
in the C-code.

Signed-off-by: Hunor Csomortáni <csomh@redhat.com>