run-prep: override __scm_apply_patch
This is the easier way of doing what we need. We need people to apply
the way they want and instead let us handle applying by patch and git.

We do this, initially, just by replacing __scm_apply_patch with our
implementation which is the same except:
1. it handles patch name and ID for sake of patch metadata
2. uses our packitpatch shell script

packitpatch is also simpler now
- it creates a git repo if .git is not present
- accepts patch name and ID (commit message, patch metadata)

Signed-off-by: Tomas Tomecek <ttomecek@redhat.com>