ensure %auto: %patch is also good
Since we override certain macros to be sure that we have a git repo in
BUILD after we're done with -bp, we need to make sure that those
overriden macros are actually executed. We also override %patch and this
change is adding %patch to the list of macros which are fine.

If %patch is being used, we don't need to rewrite %setup to %autosetup.

With this change we are able to create a source-git repo for gcc.

Workarounds: https://bugzilla.redhat.com/show_bug.cgi?id=1881840

Signed-off-by: Tomas Tomecek <ttomecek@redhat.com>