create a single-commit source-git from very hard packages
This is simple: we now have a static list of packages
VERY_VERY_HARD_PACKAGES which are "very very hard" and don't pass our
magic %prep expansion.

The solution is to run %prep the same way (to benefit from packitpatch),
discard the generated .git repo and then create a repo (or an update)
with just a single commit: all source and patches copied.

It's pretty ugly (patch files copied, not as commits) but it's the best
we can: some repo > no repo

Signed-off-by: Tomas Tomecek <ttomecek@redhat.com>
Co-authored-by: lachmanfrantisek <lachmanfrantisek@gmail.com>