worker: Apply a conversion tag in source-git
In order to be able to track which dist-git commit did source-git update
originate from, create a "conversion" tag from the dist-git branch name
and commit hash (ex: `convert/c8s/05262ab9019701b9d278b305017861200204f559`)
and apply it at the tip of the branch in source-git.

Before triggering the conversion, check if such a tag exists, and skip
updating the source-git repo if it does.

Signed-off-by: Hunor Csomortáni <csomh@redhat.com>