Use tags unique to the branch for marking the upstream boundary
This is required, because the tag marking the upstream boundary (the
"upstream ref", in Packit terms), is not allowed to be moved in CentOS
git in a non-linear fashion. This means, we are not able to reuse the
existing tag when creating new branches. Using a tag unique to each
branch solves this issue.

Signed-off-by: Hunor Csomortáni <csomh@redhat.com>