Collect metric when dist2src finished checking source-git updates
Sentry does notify us about errors, but those errors sometimes break the
loop of checking for source-git updates and we have no easy way to tell
that this happened.

Collect a metric when the loop ends, so that we can set up an alert if
the counter misses a day.

Signed-off-by: Hunor Csomortáni <csomh@redhat.com>