Don't fail if the update tag already exists
Updating a source-git repo might take some time, so a competing Celery
task (created due to some other trigger) might finish it sooner, which
might cause the update tag already be present by the time the update
process gets to the point to create it.

Expect this to happen and don't fail.

Signed-off-by: Hunor Csomortáni <csomh@redhat.com>