worker: Don't retry on 404
Retrying on 404 status codes was introduced, b/c git.centos.org often
returns 404 for `api/0/-/whoami` endpoints for an unknown reason.

But retrying on 404 for all endpoints is problematic. It causes for
example `project.exists()` calls to err with RetryError, b/c under the
hood the value returned by `exists()` relies on the return code.

Do not include 404 in the list of status codes to be retried anymore in
order to make `project.exists()` work again. Revisit the `whoami` issue
at a later date.

Signed-off-by: Hunor Csomortáni <csomh@redhat.com>